Compute S-map coefficients for a community

Compute S-map models for each time series in the block and save out the coefficients. The coefficients represent the local linear model and can be used to infer properties of the system dynamics.

compute_smap_coeffs(block, ccm_links, rescale = TRUE,
  rolling_forecast = FALSE, id_var = NULL)

Arguments

block
a data.frame containing time series for the community. Each column is a time series of abundances.

ccm_links
A data.frame containing the significant causal links. Each row is a causal link. The columns are:

xmap_from

the column index of the predictor variable in CCM

xmap_to

the column index of the predicted variable in CCM

best_E

the best embedding dimension for CCM

rescale
A logical, indicating whether to rescale each time series

rolling_forecast
A logical, indicating whether to make individual rolling forecasts for the second half of the time series.

id_var
the name of the index variable of block that is useful for labeling the time-varying output, if it exists

Value

A list with the matrix smap-coefficients for each predictor variable identified in CCM (these are the affected variables). The names in the list and the column names of the matrices use the variable names in the block.

Details

Suppose that there are causal links as follows: x --> y x --> z where --> indicates "cross-maps to". Then the interpretation from CCM is that x is affected by causes y and z. Thus, the predictive model for x should include y and z as predictors.

The S-map model is then setup as x_t+1 = F(x_t, y_t, z_t, x_t-1, x_t-2, ...) where the number of predictors is equal to the best embedding dimension for x.