plot_time_series

plot the time series in `block`, using the appropriate rescaling of the data

plot_time_series(block, time_column = censusdate, scale = "unif",
  base_size = 11, base_family = "Helvetica", base_line_size = 1)

Arguments

block
A data.frame containing time series for the community. Each column is a time series of abundances.

time_column
The name of the column in the block that has the time, which could be a numeric or a date/time type

scale
How to scale the time series: `unif` -- scale each time series to be on [0, 1] `norm` -- scale each time series to have mean 0 and variance 1 (anything else) -- no scaling

base_size
base font size

base_family
base font family

base_line_size
base size for line elements

Value

A ggplot object of the time series plot